New routes and ui styles

+++ packages/cardhost/app/components/card-renderer.js
@@ -2,16 +2,11 @@ import Component from '@glimmer/component';
 import { tracked } from '@glimmer/tracking';
 import { action } from '@ember/object';
 import { inject as service } from '@ember/service';
-// import move from 'ember-animated/motions/move';
-// import resize from 'ember-animated/motions/resize';
-// import adjustCSS from 'ember-animated/motions/adjust-css';
 import { task, waitForProperty, timeout } from 'ember-concurrency';
 import { scheduleOnce } from '@ember/runloop';
 import difference from 'lodash/difference';
 import isEqual from 'lodash/isEqual';
 import { set } from '@ember/object';
-// import { easeInAndOut } from 'ember-animated/easings/cosine';
-// import { parallel } from 'ember-animated';
 import ENV from '@cardstack/cardhost/config/environment';
 import { reads } from '@ember/object/computed';
 
@@ -30,7 +25,6 @@ export default class CardRenderer extends Component {
   @tracked mode;
   @tracked isEditMode = this.mode === 'edit';
   @tracked isViewMode = this.mode === 'view' || this.mode === 'layout';
-  @tracked isViewOrEditMode = this.isViewMode || this.mode === 'edit';
   @tracked css;
   @tracked fieldsToRemove = new Map();
   @tracked cardFocused = () => {};
@@ -203,6 +197,9 @@ export default class CardRenderer extends Component {
 
   @action
   removeCardCss() {
+    if (!this.isViewMode) {
+      return;
+    }
     this.cssService.removeCard(this.args.card, this.args.format);
   }
 
@@ -232,22 +229,4 @@ export default class CardRenderer extends Component {
       this.args.setIsSelected(value);
     }
   }
-
-  // *cardTransition({ keptSprites, receivedSprites }) {
-  //   if (receivedSprites.length) {
-  //     receivedSprites.forEach(sprite => {
-  //       sprite.applyStyles({ 'z-index': '16' });
-  //       parallel(move(sprite, { easing: easeInAndOut, duration }), resize(sprite, { easing: easeInAndOut, duration }));
-  //       adjustCSS('border-top-right-radius', sprite, { duration });
-  //       adjustCSS('border-top-left-radius', sprite, { duration });
-  //     });
-  //   }
-
-  //   keptSprites.forEach(sprite => {
-  //     sprite.applyStyles({ 'z-index': '16' });
-  //     parallel(move(sprite, { easing: easeInAndOut, duration }), resize(sprite, { easing: easeInAndOut, duration }));
-  //     adjustCSS('border-top-right-radius', sprite, { duration });
-  //     adjustCSS('border-top-left-radius', sprite, { duration });
-  //   });
-  // }
 }
